+++ addon/mixins/regenerated/models/i-c-s-soft-s-t-o-r-m-n-e-t-security-session.js
@@ -1,0 +1,11 @@
+import Ember from 'ember';
+import DS from 'ember-data';
+
+export let Model = Ember.Mixin.create({
+  userKey: DS.attr('string'),
+  startedAt: DS.attr('date'),
+  lastAccess: DS.attr('date'),
+  closed: DS.attr('boolean'),
+  validations: {
+  },
+});