+++ addon/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type.js
@@ -1,0 +1,4 @@
+import { Serializer as ObjectTypeSerializer } from '../mixins/regenerated/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type';
+import ODataSerializer from './odata';
+
+export default ODataSerializer.extend(ObjectTypeSerializer);