+++ addon/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-security-session.js
@@ -1,0 +1,4 @@
+import { Serializer as SessionSerializer } from '../mixins/regenerated/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-security-session';
+import ODataSerializer from './odata';
+
+export default ODataSerializer.extend(SessionSerializer);