+++ app/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-security-session.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-data/serializers/i-c-s-soft-s-t-o-r-m-n-e-t-security-session';