+++ tests/dummy/app/models/ember-flexberry-dummy-application-user.js
@@ -14,7 +14,7 @@ var Model = Projection.Model.extend({
   birthday: DS.attr('date'),
   gender: DS.attr('ember-flexberry-dummy-gender'),
   vip: DS.attr('boolean'),
-  karma: DS.attr('number'),
+  karma: DS.attr('decimal'),
 
   // Model validation rules.
   validations: {