+++ tests/unit/models/i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type-test.js
@@ -1,0 +1,11 @@
+import { moduleForModel, test } from 'ember-qunit';
+
+moduleForModel('i-c-s-soft-s-t-o-r-m-n-e-t-business-audit-objects-object-type', 'Unit | Model | object-type', {
+  needs: [
+  ],
+});
+
+test('it exists', function(assert) {
+  let model = this.subject();
+  assert.ok(!!model);
+});